修改销售数量的数据类型

+++ m2c.scm.application/src/main/java/cn/m2c/scm/application/order/data/bean/OrderDetailBean.java
@@ -15,7 +15,7 @@ public class OrderDetailBean {
 	@ColumnAlias(value= "goods_icon")
 	private String goodsIcon;
 	@ColumnAlias(value= "sell_num")
-	private Float  sellNum;
+	private Integer  sellNum;
 	public String getDealerOrderId() {
 		return dealerOrderId;
 	}
@@ -46,10 +46,10 @@ public class OrderDetailBean {
 	public void setGoodsIcon(String goodsIcon) {
 		this.goodsIcon = goodsIcon;
 	}
-	public Float getSellNum() {
+	public Integer getSellNum() {
 		return sellNum;
 	}
-	public void setSellNum(Float sellNum) {
+	public void setSellNum(Integer sellNum) {
 		this.sellNum = sellNum;
 	}
